protocol: speak the wire format the guest driver actually sends
The Rust half and the guest driver had never agreed. protocol/messages.rs
has one commit in its history -- the repo restructure -- and described a
protocol nobody implemented: a header of {msg_type, pad, cookie} against
the driver's {msg_type, handle, status, padding}, an open request of
{kind, index} against the driver's flat device_type, and no message at
all for four of the seven the driver sends.

Both halves compiled and a guest got as far as creating its device nodes,
so the disagreement only showed up as a backend log of "unknown msg_type
6", "unknown msg_type 7" and "invalid device kind 255" while nvidia-smi
reported a bare "Failed to initialize NVML: Unknown Error".

The driver is the authority here and not by preference: it came over from
libkrunfw complete, with the RM_CONTROL intercepts, the V1->V2 rewrite
table and the class tables, and it is the half compiled into a guest
kernel that cannot negotiate.

So the Rust side moves:

  - MsgHeader carries the handle, and status is signed. The driver tests
    (s32)status < 0 and returns it straight out of the syscall, so an
    unsigned error code stored there reads back as success and userspace
    proceeds on a failed call. Errors are written negated.
  - device_type is one flat u32: a GPU is its own minor number and the
    singletons take values above every possible minor, so decoding is a
    range check. Render nodes are refused rather than forwarded.
  - GetProcFiles and GetSysFiles are implemented. Their response is a
    bare stream of FileEntry records with no MsgHeader -- the driver
    reads from the first byte, so prefixing a header shifts everything
    and it decodes the header as a length.
  - Mmap and Munmap are answered instead of being unknown message types.

Two things found while reconciling, both real:

NvidiaBackend::drop was an inherent method rather than a Drop impl, so it
never ran: a backend that went out of scope without teardown() leaked
every host fd it held. cargo reported it only as an unused-method
warning, and the test covering it passed because dropping did nothing
observable.

IoctlReq::nested_offset is the offset of the nested block within the
payload -- always equal to data_len, since the driver lays the block
immediately after the top-level struct -- and not the offset of the
pointer field being replaced, which is what the name suggests. The
comment said the latter.

The mapping tests here still pass on a machine with no GPU by skipping,
so this is verified as far as the type system and the driver's own
structs go, and no further.

+++ device/src/error.rs
@@ -17,10 +17,10 @@ pub enum DeviceError {
     UnknownMsgType(u32),
 
     #[error("invalid device kind {0}")]
-    InvalidDeviceKind(u8),
+    InvalidDeviceKind(u32),
 
     #[error("GPU index {0} out of range")]
-    GpuIndexOutOfRange(u8),
+    GpuIndexOutOfRange(u32),
 }
 
 pub type Result<T> = std::result::Result<T, DeviceError>;
